requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
langfuse
python-dotenv
sentence-transformers